+++ packages/article/article.js
@@ -49,7 +49,6 @@ class ArticlePage extends React.Component {
           byline={byline}
           publishedTime={publishedTime}
           publicationName={publicationName}
-          style={[styles.articleMainContentRow]}
         />
       );
     } else if (rowData.type === "articleBodyRow") {