+++ packages/article/article.stories.js
@@ -1,5 +1,5 @@
 /* eslint-env browser */
-import { storiesOf } from "@storybook/react-native";
+import { storiesOf } from "dextrose/storiesOfOverloader";
 import React from "react";
 import { Platform } from "react-native";
 
@@ -8,7 +8,6 @@ import { MockedProvider } from "@times-components/utils/graphql";
 // eslint-disable-next-line import/no-unresolved
 import { addTypenameToDocument } from "apollo-utilities";
 import { query as articleQuery } from "@times-components/provider/article-provider";
-
 import Article from "./article";
 
 const fullArticleTypenameFixture = require("./fixtures/full-article-typename.json");
@@ -102,7 +101,7 @@ storiesOf("Article", module)
       );
     }
 
-    return <Article {...fullArticleFixture.data} />;
+    return <Article {...fullArticleFixture.data} analyticsStream={() => {}} />;
   })
   .add("Fixtures - No ads", () => (
     <Article {...articleFixtureNoAds.data} analyticsStream={() => {}} />