+++ packages/article/styles/meta/shared.js
@@ -4,7 +4,11 @@ const sharedStyles = {
   articleMeta: {
     marginBottom: 20,
     borderBottomColor: "#d0cece",
-    borderBottomWidth: StyleSheet.hairlineWidth
+    borderBottomWidth: StyleSheet.hairlineWidth,
+    paddingLeft: 0,
+    paddingRight: 0,
+    marginLeft: 10,
+    marginRight: 10
   },
   articleMiddleContainer: {
     paddingTop: 9