+++ packages/author-head/author-head.stories.js
@@ -1,9 +1,9 @@
 import React from "react";
-import { storiesOf } from "@storybook/react-native";
+import { storiesOf } from "dextrose/storiesOfOverloader";
 import { decorateAction, action } from "@storybook/addon-actions";
 import { withTrackingContext } from "@times-components/tracking";
+import { LateralSpacingDecorator } from "@times-components/storybook/decorators";
 import AuthorHead from "./author-head";
-import LateralSpacingDecorator from "../../storybook/decorators/lateral-spacing";
 
 const data = require("./fixtures/profile.json");