+++ packages/date-publication/__tests__/date-publication.test.js
@@ -3,7 +3,12 @@
 import "react-native";
 import React from "react";
 import renderer from "react-test-renderer";
+import Enzyme, { shallow } from "enzyme";
+import React16Adapter from "enzyme-adapter-react-16";
+import format from "date-fns/format";
 import DatePublication from "../date-publication";
+
+Enzyme.configure({ adapter: new React16Adapter() });
 
 describe("Date Publication test", () => {
   const props = {
@@ -11,8 +16,32 @@ describe("Date Publication test", () => {
     publication: "TIMES"
   };
 
-  it("renders a DatePublication component with full content", () => {
+  it("renders a DatePublication component with Times publication and Relevant date", () => {
     const tree = renderer.create(<DatePublication {...props} />).toJSON();
     expect(tree).toMatchSnapshot();
+  });
+
+  it("renders a DatePublication component with Sunday Times publication and Relevant date", () => {
+    const sundayTimesProps = {
+      ...props,
+      publication: "SUNDAYTIMES"
+    };
+    const tree = renderer
+      .create(<DatePublication {...sundayTimesProps} />)
+      .toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+
+  it("date should follow the correct format", () => {
+    const component = shallow(<DatePublication {...props} />);
+    expect(component.text()).toContain(format(props.date, "dddd MMMM DD YYYY"));
+  });
+
+  it("without providing a publication, The Times will be set as default", () => {
+    const removePubProps = props;
+    delete removePubProps.publication;
+
+    const component = shallow(<DatePublication {...removePubProps} />);
+    expect(component.text()).toContain("The Times");
   });
 });