+++ packages/image/image.stories.js
@@ -1,7 +1,7 @@
 /* eslint-disable react/no-danger */
 import React from "react";
 import { View } from "react-native";
-import { storiesOf } from "@storybook/react-native";
+import { storiesOf } from "dextrose/storiesOfOverloader";
 import Image from "./image";
 
 const squareUri =