+++ packages/link/link.stories.js
@@ -1,9 +1,9 @@
 import { View } from "react-native";
 import React from "react";
-import { storiesOf } from "@storybook/react-native";
+import { storiesOf } from "dextrose/storiesOfOverloader";
 import { action } from "@storybook/addon-actions";
+import { CenteredDecorator } from "@times-components/storybook/decorators";
 import Link, { TextLink } from "./link";
-import CenteredDecorator from "../../storybook/decorators/centered";
 
 const BigPinkSquare = props => (
   <View