+++ packages/article-summary/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.12.0"></a>
+# [0.12.0](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.11.0...@times-components/article-summary@0.12.0) (2017-12-13)
+
+
+### Features
+
+* remove font anti-aliasing of article headline on web ([#466](https://github.com/newsuk/times-components/issues/466)) ([6be8cbd](https://github.com/newsuk/times-components/commit/6be8cbd))
+
+
+
+
+<a name="0.11.0"></a>
+# [0.11.0](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.10.2...@times-components/article-summary@0.11.0) (2017-12-12)
+
+
+### Bug Fixes
+
+* resolve article tech debt ([#446](https://github.com/newsuk/times-components/issues/446)) ([4ff6a6b](https://github.com/newsuk/times-components/commit/4ff6a6b))
+
+
+### Features
+
+* refactor storybook package ([#453](https://github.com/newsuk/times-components/issues/453)) ([acadef3](https://github.com/newsuk/times-components/commit/acadef3))
+
+
+
+
+<a name="0.10.2"></a>
+## [0.10.2](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.10.1...@times-components/article-summary@0.10.2) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-summary
+
+<a name="0.10.1"></a>
+## [0.10.1](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.10.0...@times-components/article-summary@0.10.1) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/article-summary
+
 <a name="0.10.0"></a>
 # [0.10.0](https://github.com/newsuk/times-components/compare/@times-components/article-summary@0.9.2...@times-components/article-summary@0.10.0) (2017-12-07)