+++ packages/error-view/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.7.0"></a>
+# [0.7.0](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.6.4...@times-components/error-view@0.7.0) (2017-12-12)
+
+
+### Features
+
+* refactor storybook package ([#453](https://github.com/newsuk/times-components/issues/453)) ([acadef3](https://github.com/newsuk/times-components/commit/acadef3))
+
+
+
+
+<a name="0.6.4"></a>
+## [0.6.4](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.6.3...@times-components/error-view@0.6.4) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/error-view
+
+<a name="0.6.3"></a>
+## [0.6.3](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.6.2...@times-components/error-view@0.6.3) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/error-view
+
 <a name="0.6.2"></a>
 ## [0.6.2](https://github.com/newsuk/times-components/compare/@times-components/error-view@0.6.0...@times-components/error-view@0.6.2) (2017-12-04)