+++ packages/link/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.12.0"></a>
+# [0.12.0](https://github.com/newsuk/times-components/compare/@times-components/link@0.11.4...@times-components/link@0.12.0) (2017-12-12)
+
+
+### Features
+
+* refactor storybook package ([#453](https://github.com/newsuk/times-components/issues/453)) ([acadef3](https://github.com/newsuk/times-components/commit/acadef3))
+
+
+
+
+<a name="0.11.4"></a>
+## [0.11.4](https://github.com/newsuk/times-components/compare/@times-components/link@0.11.3...@times-components/link@0.11.4) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/link
+
+<a name="0.11.3"></a>
+## [0.11.3](https://github.com/newsuk/times-components/compare/@times-components/link@0.11.2...@times-components/link@0.11.3) (2017-12-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/link
+
 <a name="0.11.2"></a>
 ## [0.11.2](https://github.com/newsuk/times-components/compare/@times-components/link@0.11.0...@times-components/link@0.11.2) (2017-12-04)